+++ app-frontend/src/app/pages/admin/platform/organizations/detail/detail.module.js
@@ -1,0 +1,14 @@
+import angular from 'angular';
+
+class PlatformOrganizationDetailController {
+    constructor() {
+
+    }
+}
+
+const PlatformOrganizationDetailModule = angular.module('pages.platform.organizations.detail', []);
+PlatformOrganizationDetailModule.controller(
+    'PlatformOrganizationDetailController', PlatformOrganizationDetailController
+);
+
+export default PlatformOrganizationDetailModule;