+++ app-frontend/src/app/pages/admin/platform/organizations/detail/features/features.js
@@ -1,0 +1,13 @@
+import angular from 'angular';
+
+class AdminDetailFeaturesController {
+}
+
+const AdminDetailFeaturesModule = angular.module(
+    'pages.platform.organizations.detail.features', []
+);
+AdminDetailFeaturesModule.controller(
+    'AdminDetailFeaturesController', AdminDetailFeaturesController
+);
+
+export default AdminDetailFeaturesModule;