+++ app-frontend/src/app/pages/admin/platform/organizations/detail/settings/settings.js
@@ -1,0 +1,14 @@
+
+import angular from 'angular';
+
+class AdminDetailSettingsController {
+}
+
+const AdminDetailSettingsModule = angular.module(
+    'pages.platform.organizations.detail.settings', []
+);
+AdminDetailSettingsModule.controller(
+    'AdminDetailSettingsController', AdminDetailSettingsController
+);
+
+export default AdminDetailSettingsModule;