+++ app-frontend/src/app/pages/projects/edit/edit.module.js
@@ -2,6 +2,7 @@ import angular from 'angular';
 import {Map, Set} from 'immutable';
 import ProjectActions from '_redux/actions/project-actions';
 import AnnotationActions from '_redux/actions/annotation-actions';
+import _ from 'lodash';
 
 class ProjectsEditController {
     constructor( // eslint-disable-line max-params
@@ -55,9 +56,7 @@ class ProjectsEditController {
                         this.projectUpdateListeners.forEach((wait) => {
                             wait.resolve(project);
                         });
-                        this.getSceneList().then(() => {
-                            this.getDatasources();
-                        });
+                        this.getAndReorderSceneList();
                         if (this.project.isAOIProject) {
                             this.getPendingSceneList();
                         }
@@ -84,6 +83,21 @@ class ProjectsEditController {
                 this.fitProjectExtent();
             }
         }
+    }
+
+    getAndReorderSceneList() {
+        this.projectService.getSceneOrder(this.projectId).then(
+            (res) => {
+                this.orderedSceneId = res.results;
+            },
+            () => {
+                this.$log.log('error getting ordered scene IDs');
+            }
+        ).finally(() => {
+            this.getSceneList().then(() => {
+                this.getDatasources();
+            });
+        });
     }
 
     waitForProject() {
@@ -114,7 +128,9 @@ class ProjectsEditController {
                     (scene) => scene.statusFields.ingestStatus !== 'INGESTED'
                 ));
 
-                this.sceneList = allScenes;
+                this.sceneList = _.uniqBy(this.orderedSceneId.map((id) => {
+                    return allScenes.filter(s => s.id === id)[0];
+                }), 'id');
                 for (const scene of this.sceneList) {
                     let scenelayer = this.layerService.layerFromScene(scene, this.projectId);
                     this.sceneLayers = this.sceneLayers.set(scene.id, scenelayer);