+++ app-frontend/src/app/pages/projects/edit/scenes/scenes.module.js
@@ -1,11 +1,12 @@
+/* globals document */
 import angular from 'angular';
+
 class ProjectsScenesController {
     constructor( // eslint-disable-line max-params
         $log, $state, $scope, modalService, projectService, RasterFoundryRepository
     ) {
         'ngInject';
         this.$log = $log;
-        this.$state = $state;
         this.modalService = modalService;
         this.projectId = $state.params.projectid;
         this.$parent = $scope.$parent.$ctrl;
@@ -14,6 +15,34 @@ class ProjectsScenesController {
             name: 'Raster Foundry',
             service: RasterFoundryRepository
         };
+    }
+
+    $onInit() {
+        // eslint-disable-next-line
+        let thisItem = this;
+        this.treeOptions = {
+            dragStart: function (e) {
+                thisItem.onSceneDragStart(e);
+            },
+            dropped: function (e) {
+                thisItem.onSceneDropped(e.source.nodesScope.$modelValue);
+            }
+        };
+    }
+
+    onSceneDragStart(e) {
+        this.setDragPlaceholderHeight(e);
+    }
+
+    setDragPlaceholderHeight(e) {
+        const ele = angular.element(document.querySelector('.list-group-item'));
+        const placeholder = angular.element(e.elements.placeholder);
+        placeholder.css('height', ele.css('height'));
+    }
+
+    onSceneDropped(orderedScenes) {
+        let orderedSceneIds = orderedScenes.map(s => s.id);
+        this.updateSceneOrder(orderedSceneIds);
     }
 
     removeSceneFromProject(scene, $event) {
@@ -30,19 +59,6 @@ class ProjectsScenesController {
         );
     }
 
-    openSceneDetailModal(scene) {
-        this.$parent.removeHoveredScene();
-
-        this.modalService.open({
-            component: 'rfSceneDetailModal',
-            resolve: {
-                scene: () => scene,
-                repository: () => this.repository
-            }
-        });
-    }
-
-
     openImportModal() {
         this.modalService.open({
             component: 'rfSceneImportModal',
@@ -51,9 +67,15 @@ class ProjectsScenesController {
             }
         });
     }
+
+    updateSceneOrder(orderedSceneIds) {
+        this.projectService.updateSceneOrder(this.projectId, orderedSceneIds).then(() => {
+            this.$parent.layerFromProject();
+        });
+    }
 }
 
-const ProjectsScenesModule = angular.module('pages.projects.edit.scenes', []);
+const ProjectsScenesModule = angular.module('pages.projects.edit.scenes', ['ui.tree']);
 
 ProjectsScenesModule.controller(
     'ProjectsScenesController', ProjectsScenesController