+++ app-frontend/src/app/pages/share/share.controller.js
@@ -1,5 +1,7 @@
-import logoAsset from '../../../assets/images/logo-raster-foundry.png';
-/* global L */
+/* global BUILDCONFIG, L */
+const ASSETLOGO = BUILDCONFIG.LOGOURL || BUILDCONFIG.LOGOFILE ?
+    require(`../../../assets/images/${BUILDCONFIG.LOGOFILE}`) :
+    require('../../../assets/images/raster-foundry-logo.svg');
 
 export default class ShareController {
     constructor( // eslint-disable-line max-params
@@ -8,7 +10,7 @@ export default class ShareController {
         'ngInject';
         this.$log = $log;
         this.$state = $state;
-        this.logoAsset = logoAsset;
+        this.logoAsset = ASSETLOGO;
         this.authService = authService;
         this.projectService = projectService;
         this.mapUtilsService = mapUtilsService;